+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added the CLI options for config override: `--ambiguous-width`, `--spacing-alphabets`,
+  `--spacing-digits`, and `--spacing-punctuation-as-fullwidth`.
+
 ### Miscellaneous
 
 - Refactored the space problem checker to work on the concrete syntax tree (CST) rather than
@@ -29,8 +34,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     Possible values are `require`, `prohibit`, and `ignore`.
   - `spacing.digits` controls how to handle spaces between full-width and half-width digits.
     Possible values are the same as for alphabets.
-  - `spacing.punctuation_as_fullwidth` controls whether to treat full-width punctuation as full-width
-    characters or not.
 
 ### Fix